requirements.txt
# Core system
python-dotenv==1.0.1       # For loading environment variables from .env
psycopg2-binary==2.9.9     # PostgreSQL connector (binary version for ease)

# Numpy for vector math
numpy==1.26.4              # Latest stable version (as of now) with good compatibility

# Gemini API (Generative AI)
google-generativeai==0.5.2 # Latest known stable version for Gemini usage

# Document processing
python-docx==1.1.0         # For reading .docx files
PyMuPDF==1.23.25           # `fitz` comes from this, used for PDF processing

# Optional: regex is built-in; no need to list re
